Update frozen_dict.py
Made the resolving of hash collisions more robust.  When I realized that sets to not iterate in a consistent order, I abandoned the sorting method.  Items are still sorted by hash value, and hash collisions are still rare.  About 1 collision for every 1400 items.  When there is a hash collision, the corresponding items are put into a subclass of frozenset.